// src/types/profile.ts
export type GithubAccount = {
  following: number;
  followers: number;
  repository: number;
  url: string;
};

export type Spotify = {
  followers: number;
  playlist: number;
  url: string;
  top_tracks: { name: string; artist: string };
  currently_playing: {
    name: string;
    artist: string;
    album: string;
    id: string;
  } | null;
  last_played_track: {
    name: string;
    artist: string;
    album: string;
    id: string;
  } | null;
};

export type GitlabAccount = {
  followers: number;
  repository: number;
  url: string;
};

export type DribbbleAccount = {
  shots: number;
  liked_shots: number;
  followers: number;
  url: string;
};

export type StackOverflowAccount = {
  upvote: number;
  answers: number;
  questions: number;
  reputation: number;
  url: string;
};

export type ConnectedAccount =
  | { github: GithubAccount }
  | { gitlab: GitlabAccount }
  | { dribble: DribbbleAccount }
  | { stack_overflow: StackOverflowAccount }
  | { spotify: Spotify }
  | { linkedin_oidc: {} };